+++ examples/example.rb
@@ -1,4 +1,4 @@
-require 'radspberry'
+require_relative '../lib/radspberry'
 include DSP
 
 puts "starting simple oscillator"
@@ -17,14 +17,17 @@ sleep 1
 puts "starting crossfader (supersaw with rpmnoise)"
 chain = XFader[ o1=SuperSaw.new, o2=RpmNoise.new ]
 Speaker[ chain ]
-o1.spread  = 0.8
+o1.spread  = 0
 chain.fade = 0
-sleep 5
+puts "animating spread from 0 to 1"
+10.times do
+  o1.spread += 0.1
+  sleep 0.5
+end
+puts "crossfading with noise from 0 to 1"
 10.times do
   chain.fade += 0.1
   sleep 0.5
 end
 
-puts "muting"
 Speaker.mute
-sleep 1